refactor(calendar): unify auth and calendar into layered feature-based architecture (#12)
