+++ src/app/features/graphics/components/graphics-view/graphics-view.ts
@@ -7,7 +7,7 @@ import { VehicleService } from '../../../vehicle/services/vehicle-service/vehicl
 import { VehicleUsageHoursChartComponent } from "../vehicle-usage-hours-chart/vehicle-usage-hours-chart";
 import { MostUsedVehicleChartComponent } from "../most-used-vehicle-chart/most-used-vehicle-chart";
 import { HoursByWeekdayVehicleChartComponent } from "../hours-by-weekday-vehicle-chart/hours-by-weekday-vehicle-chart";
-import { EventService } from '../../../calendar/services/event-service/event-service';
+import { EventService } from '../../../calendar/data-access/event-service';
 
 @Component({
   selector: 'app-graphics-view',
