+++ src/styles/tokens/_typography.scss
@@ -12,5 +12,4 @@
     --font-size-lg: clamp(1.75rem, 1.2vw + 1rem, 2.25rem);
     --font-size-xl: clamp(2.25rem, 1.5vw + 1rem, 3rem);
 
-
 }
